backend/requirements.txt
fastapi
uvicorn[standard]
python-binance
google-generativeai
pydantic
python-dotenv
sqlalchemy >= 1.4 # Use a version compatible with potential async later
alembic # For database migrations
python-decouple # Alternative for config/env vars
psycopg2-binary # PostgreSQL driver
# For Async (if switching later): asyncpg, greenlet
